refactored download button
